backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg==3.2.10
psycopg2-binary==2.9.7  # Para desarrollo
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.0.0
#python-dateutil==2.8.2